Support for CimFS snapshotter
In order to use a CIM layer it needs to be mounted first. We want to share the mounted
layer CIM among different containers started from the same image. This means we can't
mount the CIM from the shim (since that is separate for each container). Hence, we need
this CimFS snapshotter. The CimFS snapshotter will take care of mounting/unmounting the
CIM when new containers are started.

Signed-off-by: Amit Barve <ambarve@microsoft.com>