Support for sharing snapshots
Contents in different namespaces can be shared by setting the content sharing policy. This
change adds the support for doing the same for snapshots. If same image is pulled and
extracted in different namespaces the same image layers will be extracted in multiple
namespaces. The snapshots created for these extracted layers contain the exact same
content. To avoid this duplication of work in improve the performance of image pull (if
the same image is already pulled in another namespace) snapshots of existing layers can be
shared across namespaces. Setting the newly added snapshot sharing policy will enable this.

Signed-off-by: Amit Barve <ambarve@microsoft.com>