Merged PR 6300726: Fix for registry spoofing
Add image pull option to force fetch blobs during image pull

This includes a new image pull option in containerd that allows the containerd client to
force containerd to fetch all image blobs during image pull even if those blobs already
are present in the content store.
This can be used when the client doesn't believe the registry and wants to forcefully
download all the content to make sure the content is altered with.

This option also forces the unpacker code to now fetch all blobs before it starts unpacking.